Tweak design of dashboard and page heading
Since the suppliers will be editing their company info back in the
Digital Marketplace it doesn't make sense to show their DUNs number
etc. in this app.

Instead this surfaces what (I think) is more important information:
Which company you are submitting services on behalf of and who you
are logged in as.

Also makes the header on the dashboard look more like the header on
the question pages.