Fix spacing between question and hint
The first few pages weren't getting the question-heading-with-hint
class applied if there was a hint in the content. We can use the
new template tag for question headings to do this automatically.

Also turns the page and question number into variables for these
pages, as I've done on later pages.