Redesign pricing input
The prototype had a pricing matrix builder which we never carried across into
the real app.

User research showed that this was both:
- Too complicated for some users to interact with
- Too simplistic for other user's pricing models

We have decided to replace it with a way of providing indicative pricing. This
commit includes the design for the new pricing input.

It does not include:
- Updated copy required for this (will be done in the content repo)
- Validation rules that understand this new input