Add error key for invalid inputs
When Play gets a response that doesn't match the type it is expecting it throws
an error. The message for this error is determined by looking up the
`validation.invalid` message, which defaults to "Incorrect value".

We can (hopefully) change this to a key which will look up the correct error
message from the content properties file.