Fix string display bug
For some text field questions, values containing quotes were being cut off when re-displayed in pages. There were two issues here: for textboxes the use of single quotes in the template meant that values with single quotes broke the template.  This is fixed by using double quotes in the templates.  For the dynamic textbox-input-list-item s there was a further problem with values being 'triple mustached' - i.e. not HTML-escaped.  Changing the script to use 'double mustacheing' instead fixes this.