Play pricing information back in the summary
The pricing information did not previously show up on the summary
page because it was stored in multiple pieces (e.g. `p8q1MinPrice`,
`p8q1Unit`).

This commit keeps the same structure but extends it by storing a
human-readable string of the price in `p8q1` which gets picked up
by the summary page.