Display all error messages on all pages
Custom error messages were not being shown on all pages.

This introduces a new method on the Fixtures class that retrieves an
error message. It first checks if the message is a key in the
content.properties file. If not, it returns the message as is. This
allows custom messages to be defined in the controller where necesarry.
This was already being done but they weren't being displayed in the
frontend.

On questions with multiple text values (eg. features and benefits), only
one error is shown per question rather than one per question value.
Because of this the validation message has to be keyed by the question
key rather than the value key.