Warn user if eSourcing ID has not been provided
If a user has not entered their eSourcing account ID in the
Digital Marketplace then we should warn them while they are
submitting their services.

This is because their services cannot be published on the
Digital Marketplace if we do not have this information to
reconcile the two systems.

The eSourcing account ID gets passed across in the cookie.
It will either be 9 numbers or not specified. If it is not
specified then the view will receive an empty string.

*So while I cannot test locally that this actually works
I am fairly confident that it will*