Return user to summary page after clicking 'edit'
When you edit a page, saving your changes should return you to
the summary page.

The default flow is still to complete one page after another.

This commit add a URL parameter (`return_to_summary`). On the
'edit' links on the summary page it is set to `yes`.

When you view a page which has `?return_to_summary=yes` after the
URL the button will say 'Save and return to summary' rather than
'Save and continue'. Clicking the button will do what the button
says it does.