Restore left orange border to required questions
Removing the `<p>` wrapping the question in the summary table
had the effect of losing the orange left border on questions
which required an answer.

Because of the design of this border it can't be applied
directly to the `<td>`. This commit adds a non-semantic `<span>`
to acheive the same effect as the `<p>`