Show feedback when service goes to/from drafts
In user research yesterday the first participant got a bit lost
trying to find the service he wanted to edit once he'd moved it
out of 'completed'.

Since we're already passing a message through to the view I thought
it might help to show it.

I want to try this out with the user research participants today,
if it works then I'll write a story and put it through QA.