Limit line length of hints and descriptions
For readability it is not recommended to have lines which span the
full width of the page. At 19px:
- The full width of the page holds ~115 characters
- ⅔ of the page width holds ~78 characters

[Dyson and Kipping (1999)](http://www.researchgate.net/publication/220106760_The_influence_of_reading_speed_and_line_length_on_the_effectiveness_of_reading_from_screen/links/09e4150ab57ed9814f000000)
found that:
- Higher reading speed was possible at 100cpl (characters per line)
  than at 55cpl
- But that comprehension was lower at 100cpl than 55cpl

GOV.UK elements recommends not exceeding 75cpl.

Setting the `max-width` of the text to 33em gives just under 70cpl.

It also aligns (not exactly, but visually close enough) with form
fields that are set to ⅔ of the page width.