Fix word counter to work with multiple instances
On loading the page the word counter only got initialised on the
first instance of a `textarea`. This meant that if there was more
than one textarea on the page it would only show the word count
when that `textarea`'s value was changed.

This commit changes the behaviour so that the word counter is
initialised for each `textarea` on page load.

Fixes [card #7905](https://www.agileplannerapp.com/boards/187120/cards/7905)