Fix erroneous validation on page 25 question 2
One of the possible responses on page 25 question 2 is
'A specific consumer group, eg Police, Defence or Health'.

This is longer than the max length validation for this question.

This commit bumps up the max length limit to allow this response
to validate.

Fixes [card #7160](https://www.agileplannerapp.com/boards/187120/cards/7160)