Do not upload if listing cannot be found
Rather than returning a 400 if the listing cannot be found just skip the
upload and return a 200.

@TheDoubleK pointed out that: "If a push task request handler returns an
HTTP status code within the range 200–299, App Engine considers the task
to have completed successfully. If the task returns a status code
outside of this range, App Engine retries the task until it succeeds."

Returning a 400 would cause the task to be retried indefinitely.
Returning a 200 prevents the task being run again.