Allow listing with "too many" completed pages to submit
Somehow one supplier's listing has been saved to the datastore with two copies of the same page of responses. By changing the "all pages have been completed" method to return true if there are *more* pages than expected as well as if there are the right number then the user will be able to mark their listing as completed.