[KK] Create new meaningful JSON format for export of services
Currently the export function uses a JSON format including the page structure
from the SSP and key names in page-question format. We want the new API to
accept JSON in a more meaningful format that properly captures what the G6
data is about. The new converter here maps existing page-question keys to
more descriptive names, and flattens the structure to get rid of the concept
of 'pages' from the data format.